pyproject.toml
[project]
name = "smart-menu"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "django (>=5.2.11,<6.0.0)",
    "django-behaviors (>=0.5.1,<0.6.0)",
    "python-decouple (>=3.8,<4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-allauth[socialaccount] (>=65.11.2,<66.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "yt-dlp[default] (>=2025.11.12,<2026.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "openai-whisper (>=20250625,<20250626)",
    "ollama (>=0.6.1,<0.7.0)",
    "faster-whisper (>=1.2.1,<2.0.0)",
    "json-repair (>=0.56.0,<0.57.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "protobuf (>=6.33.5,<7.0.0)",
    "urllib3 (>=2.6.3,<3.0.0)",
]


[tool.poetry]
packages = [
    {include = "app"},
    {include = "users"},
    {include = "product"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "django-stubs (>=5.2.5,<6.0.0)",
    "djangorestframework-stubs (>=3.16.3,<4.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "bandit (>=1.9.3,<2.0.0)",
    "pip-audit (>=2.10.0,<3.0.0)"
]


[tool.ruff]
line-length = 100
exclude = ["migrations"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing"

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.mypy]
python_version = "3.12"
strict = false
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
